+++ .github/CODEOWNERS
@@ -2,7 +2,7 @@
 /winit-android                      @MarijnS95
 
 # Apple (AppKit + UIKit)
-/src/platform/ios.rs                @madsmtm
+/winit-uikit                        @madsmtm
 /src/platform/macos.rs              @madsmtm
 /src/platform_impl/apple            @madsmtm
 /winit-common/src/core_foundation   @madsmtm
