+++ src/platform/mod.rs
@@ -5,7 +5,7 @@
 #[cfg(android_platform)]
 pub use winit_android as android;
 #[cfg(ios_platform)]
-pub mod ios;
+pub use winit_uikit as ios;
 #[cfg(macos_platform)]
 pub mod macos;
 #[cfg(orbital_platform)]
