--- src/platform_impl/apple/uikit/notification_center.rs
@@ -1,1 +1,0 @@
-../appkit/notification_center.rs
