+++ src/platform_impl/mod.rs
@@ -1,17 +1,19 @@
 #[cfg(android_platform)]
 pub(crate) use winit_android as platform;
-#[cfg(target_vendor = "apple")]
+#[cfg(macos_platform)]
 mod apple;
 #[cfg(any(x11_platform, wayland_platform))]
 mod linux;
 #[cfg(orbital_platform)]
 pub(crate) use winit_orbital as platform;
+#[cfg(ios_platform)]
+pub(crate) use winit_uikit as platform;
 #[cfg(web_platform)]
 mod web;
 #[cfg(windows_platform)]
 pub(crate) use winit_win32 as platform;
 
-#[cfg(target_vendor = "apple")]
+#[cfg(macos_platform)]
 use self::apple as platform;
 #[cfg(any(x11_platform, wayland_platform))]
 use self::linux as platform;
