+++ winit-uikit/src/notification_center.rs
@@ -1,0 +1,1 @@
+../../src/platform_impl/apple/appkit/notification_center.rs
